# Simplify code of Event.msg(msg string)