@rs 2019-07-08
Doesn't this optimize events with hooks to the expense of events with no hooks or a single hook?