@SenZhangAI 2019-07-09
Thank you ! 
I got your idea now, I was confused by this code and didn't know the reason of writing like this.

My simple test may show it may has no significant impact on optimizing events with no hooks or a single hook.  

But I don't know if it will work in other go versions or other environments, I will check the result in travis when I am not busy :)