@smyrman 2019-07-09
You can inspect the generated Asembly by issuing:
```
go build -o test .
go tool objdump test > test.s