@smyrman 2019-07-09
I am not an expert in neither Assembly nor assembly operation performance. Why do you belive the original code to be faster with no or one hook @rs?

I am asking mainly out of coriousity as I to was confused by the original code 😄