@smyrman 2019-07-12
**Disclaimer: The benchmark results and code readability is more important than the content of this comment. This is mostly a procrastination from my side.**

Again, this is _not_ my expertise, but here is my very brief assessment on the performance based on the Assembly.

# Brief analysis

The `if` causes an _conditional jump_ (`CMPQ`, `JLE`), as seen in the original code:
```asm
  event.go:130		0x40f14			4883784800		CMPQ $0x0, 0x48(AX)		
  event.go:130		0x40f19			0f8ed3000000		JLE 0x40ff2
```
The for-loop causes one _unconditional jump_ (`JMP`), and one conditional jump (`CMPQ`, `JL`):
```asm
  event.go:133		0x40f84			eb17			JMP 0x40f9d
  ...
  event.go:133		0x40fed			4839c8			CMPQ CX, AX			
  event.go:133		0x40ff0			7c94			JL 0x40f86
```

This gives the following number of conditional jumps in the execution path based on number of hooks:

- 0 hooks:
  - original code: 1 conditional jump
  - new code: 1 conditional jump
- 1 hook:
  - original code: 2 conditional jump
  - new code: 2 conditional jump
- 2 hook:
  - original code: 3 conditional jumps
  - new code: 3 conditional jumps

# Reasoning

It's my general understanding that an unconditional `JMP`, is [more or less free](https://stackoverflow.com/questions/5127833/meaningful-cost-of-the-jump-instruction) on a modern CPU.

A conditional jump is a _branch condition_, which means it affects the CPU pipeline, and could have much higher potential cost. The pipeline will be pre-filled with instructions for the CPU to execute in order or in parallel. Modern CPUs have good branch prediction to cope with this and can sometimes compute multiple "branches" at once for later to discard the branch that wasn't chosen. However the more conditional branches you get, the higher is the chance of the wrong branch being predicted, and that has some cost. Another expensive operation, is to access the main memory, which we probably need to do if we want to actually call a hook. Doing so unnecessary could waste a lot of cycles.

# Conclusion

My conclusion from that is that it's hard for me to reason _why_ or _how_ the original code (at least on Mac and compiled with Go 1.13) can be faster than your code @SenZhangAI, as they have the same number of conditional branches. This basically supports your findings from the benchmarks, that there is _no significant performance difference_.

In that case, _I_ would prefer the cleaner code.

On the other hand, I might be missing something, or the generated assembly may be different on another platform. If there _is clear evidence_ that the existing code is more performant, then a comment explaining _why_ the code is written as it is should be considered.