@rs 2019-07-19
Check the all bench, you're right, it makes no difference. I guess this trick use to do something with an older version of Go.